backend/requirements.txt
# Dependencias principales
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.4.2
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.0
python-multipart==0.0.9

# Procesamiento de texto y búsqueda
nltk==3.8.1
rank-bm25==0.2.2
scikit-learn==1.3.2
numpy==1.26.3

# Procesamiento de documentos
PyMuPDF==1.23.22
python-docx==1.1.0
openpyxl==3.1.2

# IA y LLMs
openai==1.2.0

# Autenticación y seguridad
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Bases de datos
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Utilidades
pydantic-settings==2.1.0
email-validator==2.1.0.post1
httpx==0.25.0
tenacity==8.2.3         # Para reintentos

# Pruebas
pytest==7.4.2

# Web Sockets
websockets==12.0

# Procesamiento de pagos
mercadopago==2.2.0

# Redis para WebSocket
redis==4.5.5

# Jinja2 para plantillas
jinja2==3.1.2

# Servicio de correo
fastapi-mail==1.4.2

# Gunicorn para producción
gunicorn==21.2.0 